Example output (two processes on a shared machine, so this is by no means a controlled experiment):

```
$ python -m torch.distributed.c10d.perf --path /tmp/rdv --size 2 --rank 1
 concurrency     elements             p50             p90             p99        p50 xput
           1         1000        0.000070        0.000089        0.000107     54.795 MB/s
           2         1000        0.000114        0.000143        0.000174     67.086 MB/s
           4         1000        0.000204        0.000263        0.000324     74.941 MB/s
           8         1000        0.000435        0.000572        0.000745     70.214 MB/s
           1        10000        0.000103        0.000125        0.000149    370.370 MB/s
           2        10000        0.000148        0.000173        0.000197    516.129 MB/s
           4        10000        0.000259        0.000315        0.000378    588.776 MB/s
           8        10000        0.000548        0.000669        0.000799    556.522 MB/s
           1       100000        0.000437        0.000515        0.000582    872.410 MB/s
           2       100000        0.000551        0.000589        0.000670   1385.281 MB/s
           4       100000        0.000835        0.001006        0.001171   1826.484 MB/s
           8       100000        0.001642        0.001988        0.002258   1858.979 MB/s
           1      1000000        0.004259        0.007038        0.012436    895.631 MB/s
           2      1000000        0.008137        0.008695        0.009342    937.592 MB/s
           4      1000000        0.012803        0.016692        0.019160   1191.773 MB/s
           8      1000000        0.026787        0.036418        0.043551   1139.266 MB/s
```

This does not yet do multiple types and multiple tensors per call (i.e. 1 tensor per GPU). I want to add this before merging to make it a bit more complete.