@weiyangfb 2018-07-10
@pytorchbot retest this please