@ailzhang 2018-08-28
@pytorchbot retest this please